refactor(sms): a session-free SMS layer (automations Phase 0)
Groundwork for automations, and worth doing on its own.

`sendSms` resolved the business from `getUser()`, so nothing without a
session could send a text — not a cron, not a webhook, and not an automation.
That's the blocker for "when X happens, text the customer".

`src/lib/sms.ts` now takes the business id as a parameter. `sendSms` becomes
a thin wrapper that resolves WHO is sending and delegates the how, so the
existing UI path is unchanged.

It also ends a duplicate: the ClickSend HTTP call existed twice, in
actions/sms.ts and booking/notify.ts, with different error handling — the
booking copy ignored the response entirely, so a rejected message looked
successful. Both now go through one implementation that throws on the API's
error. deriveSenderId had two copies too; the shared one is the exact
implementation that was already in production, not a rewrite, so sender IDs
don't shift.

Adds smsConfigured() so the automation builder can refuse an SMS action at
save time rather than failing at 3am.
