+++ src/lib/actions/sms.ts
@@ -5,53 +5,13 @@ import { createClient } from "@/lib/supabase/server";
 import { getActiveBizId } from "@/lib/active-business";
 
 import { getUser } from "@/lib/auth";
+import { sendSmsFor } from "@/lib/sms";
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const tbl = (sb: any, name: string) => (sb as any).from(name);
 
 /** Sanitise a business name down to an 11-char alphanumeric Sender ID for AU SMS.
  *  Strips non-alphanum, prefers PascalCase abbreviation.
  *  e.g. "Crown Roofers" → "CrownRoof", "Bob's Plumbing & Gas" → "BobsPlumbng" */
-function deriveSenderId(name: string): string {
-  const cleaned = name.replace(/[^A-Za-z0-9]+/g, "");
-  return cleaned.slice(0, 11) || "Kirei";
-}
-
-interface ClickSendResponse {
-  http_code: number;
-  data?: { messages?: Array<{ message_id?: string; status?: string }> };
-  response_msg?: string;
-}
-
-/** POST a single SMS through ClickSend. Returns the provider message id +
- *  status. Throws with the API's error message on failure. */
-async function clicksendSend(args: { from: string; to: string; body: string }): Promise<{ id: string; status: string }> {
-  const username = process.env.CLICKSEND_USERNAME;
-  const apiKey   = process.env.CLICKSEND_API_KEY;
-  if (!username || !apiKey) throw new Error("ClickSend credentials not configured");
-
-  const auth = Buffer.from(`${username}:${apiKey}`).toString("base64");
-  const res = await fetch("https://rest.clicksend.com/v3/sms/send", {
-    method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-      "Authorization": `Basic ${auth}`,
-    },
-    body: JSON.stringify({
-      messages: [{ from: args.from, to: args.to, body: args.body, source: "kirei" }],
-    }),
-  });
-
-  const json = (await res.json().catch(() => ({}))) as ClickSendResponse;
-  if (!res.ok || json.http_code !== 200) {
-    throw new Error(`SMS send failed: ${json.response_msg ?? res.statusText}`);
-  }
-  const msg = json.data?.messages?.[0];
-  return {
-    id:     msg?.message_id ?? "",
-    status: msg?.status     ?? "sent",
-  };
-}
-
 export interface SmsConversation {
   id: string;
   business_id: string;
@@ -103,6 +63,13 @@ export async function getMessages(conversationId: string): Promise<SmsMessage[]>
   return (data ?? []) as SmsMessage[];
 }
 
+/**
+ * Send a text as the signed-in user's business.
+ *
+ * The work lives in lib/sms.ts, which takes the business id explicitly so
+ * crons, webhooks and automations can send too. This wrapper only resolves
+ * WHO is sending.
+ */
 export async function sendSms(payload: {
   to: string;           // E.164 phone, e.g. +61412345678
   body: string;
@@ -113,68 +80,10 @@ export async function sendSms(payload: {
   const user = await getUser();
   const businessId = await getActiveBizId(supabase, user.id);
 
-  // Resolve the Sender ID — use the business's saved sms_sender_id if set,
-  // otherwise derive an 11-char alpha tag from their name. Customers see this
-  // as the "from" on their phone instead of an unfamiliar number.
-  const { data: business } = await tbl(supabase, "businesses")
-    .select("name, sms_sender_id")
-    .eq("id", businessId)
-    .single();
-  const fromName = business?.sms_sender_id?.trim() || deriveSenderId(business?.name ?? "Kirei");
-
-  // Get or create conversation
-  let { data: conv } = await tbl(supabase, "sms_conversations")
-    .select("id")
-    .eq("business_id", businessId)
-    .eq("customer_phone", payload.to)
-    .maybeSingle();
-
-  if (!conv) {
-    const { data: newConv, error: convErr } = await tbl(supabase, "sms_conversations")
-      .insert({
-        business_id: businessId,
-        customer_id: payload.customerId ?? null,
-        customer_name: payload.customerName,
-        customer_phone: payload.to,
-        last_message_at: new Date().toISOString(),
-      })
-      .select("id")
-      .single();
-    if (convErr) throw convErr;
-    conv = newConv;
-  }
-
-  // Send via ClickSend
-  const sent = await clicksendSend({
-    from: fromName,
-    to:   payload.to,
-    body: payload.body,
-  });
-
-  // Save message to DB. We reuse the existing twilio_sid column to store the
-  // ClickSend message id — schema-compat for now, can rename later.
-  const { data: msg, error: msgErr } = await tbl(supabase, "sms_messages")
-    .insert({
-      conversation_id: conv.id,
-      business_id: businessId,
-      direction: "outbound",
-      body: payload.body,
-      from_number: fromName,
-      to_number: payload.to,
-      twilio_sid: sent.id,
-      status: sent.status,
-    })
-    .select("id")
-    .single();
-  if (msgErr) throw msgErr;
-
-  // Update conversation last_message_at
-  await tbl(supabase, "sms_conversations")
-    .update({ last_message_at: new Date().toISOString() })
-    .eq("id", conv.id);
+  const res = await sendSmsFor(supabase, businessId, payload);
 
   revalidatePath("/messages");
-  return { conversationId: conv.id, messageId: msg.id };
+  return { conversationId: res.conversationId, messageId: res.messageId };
 }
 
 export async function markConversationRead(conversationId: string): Promise<void> {
